added api_uth
moved the user manager and user model to the api_auth app .
moved the serializers and the views (login ,refres_token,user_profile)
