+++ project/api_main/serializers.py
@@ -1,6 +1,7 @@
 from rest_framework import serializers
 from api_main.models import Profile
-from api_agency.serializers import UserSerializer,AgencyDetailSerializer,VehicleDetailsSerializer,UserDetailsSerializer
+from api_agency.serializers import AgencyDetailSerializer,VehicleDetailsSerializer
+from api_auth.serializers import UserSerializer,UserDetailsSerializer
 from api_agency.models import Reservation
 
 class ProfileSerializer(serializers.ModelSerializer):
